+++ src/components/ContractDetail/AddressOverview/styles.ts
@@ -64,3 +64,16 @@ export const RedirectButton = styled(Box)(({ theme }) => ({
     backgroundColor: alpha(theme.palette.common.black, 0.8)
   }
 }));
+
+export const BannerSuccess = styled(Box)`
+  font-weight: 500;
+  font-size: 18px;
+  line-height: 21px;
+  text-align: center;
+  color: #ffffff;
+  background: #5a9c56;
+  border-radius: 10px;
+  width: 100%;
+  padding: 12px 0px;
+  margin: 20px 0px;
+`;
